# UI/UX fixes: Reduce 3d viewer loading time + fix blurry renders on high-end displays