@gregzaal 2022-11-07
Hey, thanks for the PR! I should have mentioned in the readme that the `dev` branch is the one to work on - I've already fixed the loading delay/lack of spinner issue there.

If you can submit a separate PR for the `dpr` issue built against the `dev` branch that'd be good to go :)

As for switching the preview resolution down to 2K to save time, I'd rather not do this. 4K is a good compromise between loading time and preview quality (most of our assets are at least 8K if not more).